+++ app/controllers/DiscussionsController.php
@@ -904,128 +904,6 @@ class DiscussionsController extends ControllerBase
     }
 
     /**
-     * Shows the user profile
-     *
-     * @todo Move to the UsersController
-     * @param $id
-     * @param $username
-     */
-    public function userAction($id, $username)
-    {
-        $user = $id ? Users::findFirstById($id) : Users::findFirstByLogin($username);
-        if (!$user) {
-            $user = Users::findFirstByName($username);
-        }
-
-        if (!$user) {
-            $this->flashSession->error('The user does not exist');
-            $this->response->redirect();
-            return;
-        }
-
-        $this->view->setVar('user', $user);
-
-        $parametersNumberPosts = [
-            'users_id = ?0 AND deleted = 0',
-            'bind' => [$user->id]
-        ];
-        $this->view->setVar('numberPosts', Posts::count($parametersNumberPosts));
-
-        $parametersNumberReplies = [
-            'users_id = ?0',
-            'bind' => [$user->id]
-        ];
-        $this->view->setVar('numberReplies', PostsReplies::count($parametersNumberReplies));
-
-        $parametersActivities = [
-            'users_id = ?0',
-            'bind'  => [$user->id],
-            'order' => 'created_at DESC',
-            'limit' => 15
-        ];
-        $this->view->setVar('activities', Activities::find($parametersActivities));
-
-        $users   = Users::find(['columns' => 'id', 'conditions' => 'karma != 0', 'order' => 'karma DESC']);
-        $ranking = count($users);
-        foreach ($users as $position => $everyUser) {
-            if ($everyUser->id == $user->id) {
-                $ranking = $position + 1;
-                break;
-            }
-        }
-
-        $this->view->setVars([
-            'ranking'       => $ranking,
-            'total_ranking' => count($users),
-        ]);
-
-        $this->gravatar->setSize(64);
-        $this->tag->setTitle('Profile - ' . $this->escaper->escapeHtml($user->name));
-    }
-
-    /**
-     * Allow to change your user settings
-     */
-    public function settingsAction()
-    {
-
-        $usersId = $this->session->get('identity');
-        if (!$usersId) {
-            $this->flashSession->error('You must be logged first');
-            return $this->response->redirect();
-        }
-
-        $user = Users::findFirstById($usersId);
-        if (!$user) {
-            $this->flashSession->error('The user does not exist');
-            return $this->response->redirect();
-        }
-
-        if ($this->request->isPost()) {
-            if (!$this->checkTokenPost()) {
-                return $this->response->redirect();
-            }
-
-            $user->timezone      = $this->request->getPost('timezone');
-            $user->notifications = $this->request->getPost('notifications');
-            $user->theme         = $this->request->getPost('theme');
-            $user->digest        = $this->request->getPost('digest');
-            if ($user->save()) {
-                $this->session->set('identity-theme', $user->theme);
-                $this->session->get('identity-timezone', $user->timezone);
-                $this->flashSession->success('Settings were successfully updated');
-                return $this->response->redirect();
-            }
-
-        } else {
-            $this->tag->displayTo('timezone', $user->timezone);
-            $this->tag->displayTo('notifications', $user->notifications);
-            $this->tag->displayTo('theme', $user->theme);
-            $this->tag->displayTo('digest', $user->digest);
-        }
-
-        $this->tag->setTitle('My Settings');
-        $this->tag->setAutoEscape(false);
-
-        $this->view->user      = $user;
-        $this->view->timezones = require APP_PATH .'/app/config/timezones.php';
-
-        $parametersNumberPosts = [
-            'users_id = ?0 AND deleted = 0',
-            'bind' => [$user->id]
-        ];
-        $this->view->numberPosts = Posts::count($parametersNumberPosts);
-
-        $parametersNumberReplies = [
-            'users_id = ?0',
-            'bind' => [$user->id]
-        ];
-
-        $this->gravatar->setSize(64);
-        $this->view->numberReplies = PostsReplies::count($parametersNumberReplies);
-    }
-
-    /**
      * Shows the latest notifications for the current user
      */
     public function notificationsAction($offset = 0)
