+++ app/views/layouts/users.volt
@@ -1,0 +1,5 @@
+{% include 'partials/top-menu.volt' %}
+
+{{ content() }}
+
+{% include 'partials/footer.volt' %}
